+++ src/expression/function/impl/ApplyRegex.ts
@@ -4,13 +4,17 @@ import { FuncArgSimpleRegex } from "../arg/FuncArgSimpleRegex.js";
 import { FuncArgRegex } from "../arg/FuncArgRegex.js";
 import { FuncArgBoolean } from "../arg/FuncArgBoolean.js";
 import { BooleanValue } from "../../../value/BooleanValue.js";
+import { StringValue } from "../../../value/StringValue";
+import { FuncArgRegexFlag } from "../arg/FuncArgRegexFlag";
 
 export abstract class ApplyRegex extends Func {
   protected static readonly KEY_INPUT = "input";
   protected static readonly KEY_PATTERN = "pattern";
   protected static readonly KEY_CASE_INSENSITIVE = "caseInsensitive";
+  protected static readonly KEY_FLAGS = "flags";
 
   protected readonly requireFullMatch: boolean;
+  protected readonly simplePattern: boolean;
 
   protected constructor(
     name: string,
@@ -22,12 +26,19 @@ export abstract class ApplyRegex extends Func {
       simplePattern
         ? new FuncArgSimpleRegex(true, ApplyRegex.KEY_PATTERN)
         : new FuncArgRegex(true, ApplyRegex.KEY_PATTERN),
-      new FuncArgBoolean(
-        false,
-        ApplyRegex.KEY_CASE_INSENSITIVE,
-        BooleanValue.of(false)
-      ),
+      simplePattern
+        ? new FuncArgBoolean(
+            false,
+            ApplyRegex.KEY_CASE_INSENSITIVE,
+            BooleanValue.of(false)
+          )
+        : new FuncArgRegexFlag(
+            false,
+            ApplyRegex.KEY_FLAGS,
+            new StringValue("gim")
+          ),
     ]);
     this.requireFullMatch = requireFullMatch;
+    this.simplePattern = simplePattern;
   }
 }